+++ src/main/kotlin/com/freya02/botcommands/internal/commands/application/ApplicationCommandListener.kt
@@ -8,11 +8,11 @@ import com.freya02.botcommands.api.core.service.getInterfacedServices
 import com.freya02.botcommands.api.core.utils.getMissingPermissions
 import com.freya02.botcommands.api.core.utils.simpleNestedName
 import com.freya02.botcommands.internal.BContextImpl
-import com.freya02.botcommands.internal.ExceptionHandler
 import com.freya02.botcommands.internal.Usability
 import com.freya02.botcommands.internal.Usability.UnusableReason
 import com.freya02.botcommands.internal.commands.application.slash.SlashCommandInfo
 import com.freya02.botcommands.internal.commands.ratelimit.withRateLimit
+import com.freya02.botcommands.internal.core.ExceptionHandler
 import com.freya02.botcommands.internal.utils.throwInternal
 import dev.minn.jda.ktx.messages.reply_
 import kotlinx.coroutines.launch
