+++ src/main/kotlin/com/freya02/botcommands/internal/modals/ModalListener.kt
@@ -3,7 +3,7 @@ package com.freya02.botcommands.internal.modals
 import com.freya02.botcommands.api.core.annotations.BEventListener
 import com.freya02.botcommands.api.core.service.annotations.BService
 import com.freya02.botcommands.internal.BContextImpl
-import com.freya02.botcommands.internal.ExceptionHandler
+import com.freya02.botcommands.internal.core.ExceptionHandler
 import com.freya02.botcommands.internal.utils.throwUser
 import dev.minn.jda.ktx.messages.reply_
 import dev.minn.jda.ktx.messages.send
